fix: wire vault DDL into TiDB schema init, remove db9 vault DDL
Address adversary review blocking points:

1. Vault DDL now appended to both tidbAutoEmbeddingSchemaStatements()
   and tidbAppEmbeddingSchemaStatements() via new
   schema.VaultTiDBSchemaStatements() in pkg/tenant/schema/vault.go

2. ValidateTiDBSchemaForMode() now checks all 6 vault tables exist

3. Removed PG-native vault DDL from db9/schema.go — vault is
   TiDB/MySQL-only per product decision

4. vault.SchemaStatements() now delegates to
   schema.VaultTiDBSchemaStatements() (single source of truth)

5. Updated schema_test.go to check vault tables in TiDB providers
   instead of db9
